Gotova Devetka

+++ backend/routes/posts.js
@@ -34,10 +34,11 @@ router.post(
   (req, res, next) => {
     const url = req.protocol + '://' + req.get("host");           // req.protocol vraca ili http ili https
     const post = new Post({
-     title: req.body.title,
-     content: req.body.content,
-     imagePath: url + "/images/" + req.file.filename              // file property stvara multer kao i filename
-  });
+      title: req.body.title,
+      content: req.body.content,
+      imagePath: url + "/images/" + req.file.filename,              // file property stvara multer kao i filename
+      creator: req.userData.userId
+    });
   post.save().then(createdPost => {                               // createdPost je rezultat tj. napravljeni Post()
     res.status(201).json({
       message: 'Post added succesfully',
@@ -63,10 +64,15 @@ router.put(
      _id: req.body.id,
       title: req.body.title,
       content: req.body.content,
-      imagePath: imagePath
+      imagePath: imagePath,
+      creator: req.userData.userId
   });
-  Post.updateOne({_id: req.params.id}, post).then(result => {
-    res.status(200).json({ message: 'Update successful!' })
+  Post.updateOne({ _id: req.params.id, creator: req.userData.userId }, post).then(result => {
+    if (result.nModified > 0) {
+      res.status(200).json({ message: 'Update successful!' })
+    } else {
+    res.status(401).json({ message: 'Not authorized' })
+    }
   });
 });
 
@@ -107,9 +113,12 @@ router.get("/:id", checkAuth, (req, res, next) =>{
 })
 
 router.delete("/:id", checkAuth, (req, res, next) => {
-  Post.deleteOne({_id: req.params.id}).then(result => {
-    console.log(result);
-    res.status(200).json({message: "Post deteletd"});         // req.params. Tu ide dinamicki parametar koji dodje gore posle ":" kao sto je ovde id
+  Post.deleteOne({ _id: req.params.id, creator: req.userData.userId }).then(result => { // req.params. Tu ide dinamicki parametar koji dodje gore posle ":" kao sto je ovde id
+    if (result.n > 0) {
+      res.status(200).json({ message: 'Deletion successful!' })
+    } else {
+    res.status(401).json({ message: 'Not authorized' })
+    }
   });
 });
